requirements.txt
click==8.1.7
openai==1.75.0
python-dotenv==1.1.0
